Add files via upload
A tool for capturing and managing viewport display settings in Maya. It can generate MEL scripts and presets that let you easily reuse the same settings later. These configurations can be saved to a shelf, assigned to a shortcut, or shared with other users via a network drive. This is especially useful for creating consistent-looking playblasts in previs, layout, and animation, leading to more coherent context edits and reducing viewer distraction.